chore(thumb): rename and tidy the example app
- Rename the example stream_video_thumbnail_example -> stream_thumbnail_example
  (pubspec, Android namespace/package dir, iOS bundle id).
- Trim example/pubspec.yaml to essentials (drop flutter-create boilerplate,
  cupertino_icons, and flutter_lints).
- Remove example/analysis_options.yaml so it inherits the workspace config.
