+++ packages/stream_thumbnail/example/android/app/build.gradle.kts
@@ -5,7 +5,7 @@ plugins {
 }
 
 android {
-    namespace = "io.getstream.stream_video_thumbnail_example"
+    namespace = "io.getstream.stream_thumbnail_example"
     compileSdk = flutter.compileSdkVersion
     ndkVersion = flutter.ndkVersion
 
@@ -16,7 +16,7 @@ android {
 
     defaultConfig {
         // TODO: Specify your own unique Application ID (https://developer.android.com/studio/build/application-id.html).
-        applicationId = "io.getstream.stream_video_thumbnail_example"
+        applicationId = "io.getstream.stream_thumbnail_example"
         // You can update the following values to match your application needs.
         // For more information, see: https://flutter.dev/to/review-gradle-config.
         minSdk = flutter.minSdkVersion
